pynxt/cli: fix AVR stuck in update mode after jump
When flashing the firmware, the NXT would get stuck in update mode, so
the firmware would start but the AVR would not send updates on the
current status of buttons, etc.

This fixes by waiting for the AVR to be ready for another command
before telling it to jump to the new firmware.